# api/openapi/rest.yaml
openapi: 3.0.3
info:
  title: TimeLog API
  version: 1.0.0
  description: |
    REST API for the TimeLog application. Payload DTO schemas are generated from
    `api/proto/timelog/v1`. The `{ data, message, status }` envelope is defined here.
servers:
  - url: /api
    description: API base path (relative to app origin)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                nullable: true
              message:
                type: string
              status:
                type: integer
    TimelogEnvelope:
      description: Timelog payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Timelog'
              message:
                type: string
              status:
                type: integer
    TimelogListEnvelope:
      description: Timelog list payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Timelog'
              message:
                type: string
              status:
                type: integer
    CategoryEnvelope:
      description: Category payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Category'
              message:
                type: string
              status:
                type: integer
    CategoryListEnvelope:
      description: Category list payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              message:
                type: string
              status:
                type: integer
    CategoryTreeEnvelope:
      description: Category tree payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryTreeNode'
              message:
                type: string
              status:
                type: integer
...
    NullDataEnvelope:
      description: Command succeeded with null data
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                nullable: true
              message:
                type: string
              status:
                type: integer
...
paths:
  /timelogs:
    get:
      tags: [timelog]
      summary: List time logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/TimelogListEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    post:
      tags: [timelog]
      summary: Create time log
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimelogRequest'
      responses:
        '200':
          $ref: '#/components/responses/TimelogEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /timelogs/{id}:
    get:
      tags: [timelog]
      summary: Get time log
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/TimelogEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    put:
      tags: [timelog]
      summary: Update time log
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimelogRequest'
      responses:
        '200':
          $ref: '#/components/responses/TimelogEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    delete:
      tags: [timelog]
      summary: Delete time log
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /categories:
    get:
      tags: [category]
      summary: List categories
      parameters:
        - name: level
          in: query
          schema:
            type: integer
        - name: parent_id
          in: query
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/CategoryListEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    post:
      tags: [category]
      summary: Create category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '200':
          $ref: '#/components/responses/CategoryEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /categories/tree:
    get:
      tags: [category]
      summary: Get category tree
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/CategoryTreeEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /categories/{id}:
    get:
      tags: [category]
      summary: Get category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/CategoryEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    put:
      tags: [category]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          $ref: '#/components/responses/CategoryEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /categories/{id}/move:
    post:
      tags: [category]
      summary: Move category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveCategoryRequest'
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
...